Refactor layout code out of renderer and fix conditional regions
